Add domino selection and tracking

+++ Assets/Scripts/Models/Track.cs
@@ -8,11 +8,11 @@ namespace Assets.Scripts.Models
 {
     public class Track
     {
-        public int? PlayerId { get; set; }  // each player gets only 1 track. When null then anyone can use this track
+        public ulong? PlayerId { get; set; }  // each player gets only 1 track. When null then anyone can use this track
         public bool HasTrain { get; set; }  // TODO: needs set to true when a player is unable to play a domino
         public List<int> DominoIds { get; set; }
 
-        public Track(int dominoId, int? owner = null)
+        public Track(int dominoId, ulong? owner = null)
         {
             DominoIds = new List<int>();
             DominoIds.Add(dominoId);
